Remove IRC as a recommended medium for contact
Both IRC and Gitter are not working well enough for us to handle
arriving prospects;

(a) because the mentors, developers, and reviewers are not present,

(b) it deceives students into trusting random people who have given
incorrect or incomplete advice,

The better forum is our mailing list.

This doesn't stop us from using IRC or Gitter during GSoC, of course.