Remove unverified mentors
It has been two weeks since we asked those hoping to be mentors to read
and review the project ideas page.

http://lists.sugarlabs.org/archive/sugar-devel/2020-February/057783.html

So remove all mentors who have not responded to that mail.

Reading and answering mail on sugar-devel@ should be a fundamental
requirement of being a mentor for Sugar Labs.