Adjust proposal requirements
Based on reading of GSOC 2020 mail dated 8th March 14:48:15 -0700.

- demote posting on the mailing list from MUST to SHOULD,

- clarify that posting on the mailing list is community interactions by
  a student, but that students may choose not to interact if they want
  their proposal deleted; students are in control of this.

My opinion; (On the other hand, a proposal without interaction is going
to have to be remarkably good for it to outweigh the interaction by
other students.  We are looking for students who will stick with us;
we're not looking specifically for code.)